# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "Accepted": {
        "additionalProperties": false,
        "description": "A long operation that has been accepted but not completed. AC-B9.",
        "properties": {
          "detail": {
            "description": "What was accepted.",
            "title": "Detail",
            "type": "string"
          },
          "events": {
            "description": "Where to watch this run's event stream.",
            "title": "Events",
            "type": "string"
          },
          "runId": {
            "description": "The run this operation concerns.",
            "format": "uuid",
            "title": "Runid",
            "type": "string"
          },
          "runIdentity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "SHA-256 of the specification and its resolved input artefact hashes (AC-F1). Two submissions of the same specification are two runs with one identity: the identifier says when, the identity says what.",
            "title": "Runidentity"
          },
          "status": {
            "const": "accepted",
            "description": "Always `accepted`.",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "runId",
          "status",
          "detail",
          "events"
        ],
        "title": "Accepted",
        "type": "object"
      },
...
      "CorpusOut": {
        "additionalProperties": false,
        "description": "One jurisdiction's corpus, as the curation screen shows it. S05.",
        "properties": {
          "awaiting": {
            "description": "How many are still short of CURATED.",
            "title": "Awaiting",
            "type": "integer"
          },
          "curated": {
            "description": "How many have reached CURATED.",
            "title": "Curated",
            "type": "integer"
          },
          "jurisdiction": {
            "description": "ISO 3166-1 alpha-3 code.",
            "title": "Jurisdiction",
            "type": "string"
          },
          "latestRetrieval": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the most recent source was retrieved.",
            "title": "Latestretrieval"
          },
          "licences": {
            "description": "Every distinct licence in this corpus.",
            "items": {
              "type": "string"
            },
            "title": "Licences",
            "type": "array"
          },
          "missingDpia": {
            "description": "How many declare personal data with no DPIA reference. Counted separately because it is a defect rather than a state: the register should not hold one.",
            "title": "Missingdpia",
            "type": "integer"
          },
          "personalDataSources": {
            "description": "How many declare personal data.",
            "title": "Personaldatasources",
            "type": "integer"
          },
          "quarantined": {
            "description": "How many were refused by the licence gate.",
            "title": "Quarantined",
            "type": "integer"
          },
          "sources": {
            "description": "How many sources are registered for it.",
            "title": "Sources",
            "type": "integer"
          }
        },
        "required": [
          "jurisdiction",
          "sources",
          "curated",
          "quarantined",
          "awaiting",
          "personalDataSources",
          "missingDpia",
          "licences"
        ],
        "title": "CorpusOut",
        "type": "object"
      },
      "CorpusPage": {
        "additionalProperties": false,
        "description": "The corpora at this site, by jurisdiction.",
        "properties": {
          "items": {
            "description": "One entry per jurisdiction.",
            "items": {
              "$ref": "#/components/schemas/CorpusOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "CorpusPage",
        "type": "object"
      },
...
      "Problem": {
        "description": "An RFC 9457 problem document.",
        "properties": {
          "code": {
            "description": "Machine readable, stable problem code.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explanation for this occurrence.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "URI of this occurrence.",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human readable summary.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "RetentionOut": {
        "additionalProperties": false,
        "description": "One retention action, due or executed. S06, SAD 7.3.",
        "properties": {
          "approvedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Who approved the deletion. Null means unapproved: SAD 7.3 makes deletion an approved, ledgered action rather than a timer firing.",
            "title": "Approvedby"
          },
          "daysRemaining": {
            "description": "Days until due. Negative when overdue.",
            "title": "Daysremaining",
            "type": "integer"
          },
          "dueAt": {
            "description": "When the retention period expires.",
            "format": "date-time",
            "title": "Dueat",
            "type": "string"
          },
          "executedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When it was carried out.",
            "title": "Executedat"
          },
          "id": {
            "description": "The action.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "manifestsRetained": {
            "description": "Whether the manifests survive the deletion. They must: a lineage that loses its hashes when a corpus is deleted cannot be verified afterwards.",
            "title": "Manifestsretained",
            "type": "boolean"
          },
          "policy": {
            "description": "The retention policy that scheduled it.",
            "title": "Policy",
            "type": "string"
          },
          "subject": {
            "description": "What that subject is, in words.",
            "title": "Subject",
            "type": "string"
          },
          "subjectId": {
            "description": "What it applies to.",
            "format": "uuid",
            "title": "Subjectid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subjectId",
          "subject",
          "policy",
          "dueAt",
          "manifestsRetained",
          "daysRemaining"
        ],
        "title": "RetentionOut",
        "type": "object"
      },
      "RetentionPage": {
        "additionalProperties": false,
        "description": "Retention actions at this site, soonest first.",
        "properties": {
          "items": {
            "description": "The scheduled actions.",
            "items": {
              "$ref": "#/components/schemas/RetentionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "overdue": {
            "description": "How many are past their due date and not executed.",
            "title": "Overdue",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "overdue"
        ],
        "title": "RetentionPage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nThe DRAUPNIR control plane API.\n\nErrors are RFC 9457 problem documents served as `application/problem+json`.\nMutating endpoints require an `Idempotency-Key`. Collections are cursor\npaginated. Mutable resources carry an `ETag` and require `If-Match`.\nLong operations return 202 with a run identifier; nothing blocks on training.\n",
    "summary": "CIM-56 model factory control plane",
    "title": "DRAUPNIR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/corpora": {
      "get": {
        "description": "The corpus of each jurisdiction, and how far curation has reached. S05.\n\n    The counts are what a curator acts on: how many sources are registered,\n    how many cleared the licence gate, and how many were quarantined by it. A\n    quarantined source is not a failure of the pipeline -- it is the licence\n    gate doing its job -- so it is counted beside the others rather than\n    hidden in an error state.\n    \n\nRequires: `admin`, `approver`, `curator`, `operator`, `viewer`.",
        "operationId": "listCorpora",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorpusPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Corpora by jurisdiction, with curation progress",
        "tags": [
          "corpora"
        ]
      }
    },
    "/v1/corpora/{iso3}/curate": {
      "post": {
        "description": "Deduplicate, filter and decontaminate against the evaluation sets. 202.\n\nRequires: `curator`.",
        "operationId": "curateCorpus",
        "parameters": [
          {
            "description": "ISO 3166-1 alpha-3 code of the jurisdiction.",
            "in": "path",
            "name": "iso3",
            "required": true,
            "schema": {
              "description": "ISO 3166-1 alpha-3 code of the jurisdiction.",
              "maxLength": 3,
              "minLength": 3,
              "pattern": "^[A-Z]{3}$",
              "title": "Iso3",
              "type": "string"
            }
          },
          {
            "description": "Replaying a request with the same key returns the original result.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Replaying a request with the same key returns the original result.",
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Run the curation pipeline",
        "tags": [
          "corpora"
        ]
      }
    },
    "/v1/corpora/{iso3}/ingest": {
      "post": {
        "description": "Stage, hash, publish, seal and register. Returns 202.\n\n    Hashing a corpus is minutes to hours of work; AC-B9 requires that no\n    endpoint blocks an HTTP request on it.\n    \n\nRequires: `curator`.",
        "operationId": "ingestCorpus",
        "parameters": [
          {
            "description": "ISO 3166-1 alpha-3 code of the jurisdiction.",
            "in": "path",
            "name": "iso3",
            "required": true,
            "schema": {
              "description": "ISO 3166-1 alpha-3 code of the jurisdiction.",
              "maxLength": 3,
              "minLength": 3,
              "pattern": "^[A-Z]{3}$",
              "title": "Iso3",
              "type": "string"
            }
          },
          {
            "description": "Replaying a request with the same key returns the original result.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Replaying a request with the same key returns the original result.",
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Ingest and hash a jurisdiction's sources",
        "tags": [
          "corpora"
        ]
      }
    },
...
    "/v1/retention": {
      "get": {
        "description": "Corpora approaching their deletion point. S06, SAD 7.3.\n\n    Deletion here is an approved, ledgered action rather than a timer firing,\n    so an unapproved action that is past its due date is a thing somebody has\n    to decide about rather than a thing that has silently happened. The overdue\n    count is on the response for exactly that reason.\n    \n\nRequires: `admin`, `approver`, `curator`, `operator`, `viewer`.",
        "operationId": "listRetention",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Retention actions, due and executed",
        "tags": [
          "corpora"
        ]
      }
    },
...
  }
}